test: merge existing string tests
