+++ test/Makefile.autosetup
@@ -3,7 +3,6 @@ TEST_OBJS   = test/main.o \
 	      test/md5.o \
 	      test/path.o \
 	      test/rfc2047.o \
-	      test/string.o \
 	      test/address.o \
 	      test/url.o \
 	      test/file.o
