+++ test/string/mutt_str_strnfcpy.c
@@ -41,4 +41,64 @@ void test_mutt_str_strnfcpy(void)
   {
     TEST_CHECK(mutt_str_strnfcpy(NULL, NULL, 5, 10) == 0);
   }
+
+  const char src[] = "One Two Three Four Five";
+  char dst[10];
+  char big[32];
+
+  { /* copy a substring */
+    size_t len = mutt_str_strnfcpy(dst, src, 3, sizeof(dst));
+    if (!TEST_CHECK(len == 3))
+    {
+      TEST_MSG("Expected: %zu", 3);
+      TEST_MSG("Actual  : %zu", len);
+    }
+    if (!TEST_CHECK(strcmp(dst, "One") == 0))
+    {
+      TEST_MSG("Expected: %s", "One");
+      TEST_MSG("Actual  : %s", dst);
+    }
+  }
+
+  { /* try to copy the whole available length */
+    size_t len = mutt_str_strnfcpy(dst, src, sizeof(dst), sizeof(dst));
+    if (!TEST_CHECK(len == sizeof(dst) - 1))
+    {
+      TEST_MSG("Expected: %zu", sizeof(dst) - 1);
+      TEST_MSG("Actual  : %zu", len);
+    }
+    if (!TEST_CHECK(strcmp(dst, "One Two T") == 0))
+    {
+      TEST_MSG("Expected: %s", "One Two T");
+      TEST_MSG("Actual  : %s", dst);
+    }
+  }
+
+  { /* try to copy more than it fits */
+    size_t len = mutt_str_strnfcpy(dst, src, strlen(src), sizeof(dst));
+    if (!TEST_CHECK(len == sizeof(dst) - 1))
+    {
+      TEST_MSG("Expected: %zu", sizeof(dst) - 1);
+      TEST_MSG("Actual  : %zu", len);
+    }
+    if (!TEST_CHECK(strcmp(dst, "One Two T") == 0))
+    {
+      TEST_MSG("Expected: %s", "One Two T");
+      TEST_MSG("Actual  : %s", dst);
+    }
+  }
+
+  { /* try to copy more than available */
+    size_t len = mutt_str_strnfcpy(big, src, sizeof(big), sizeof(big));
+    if (!TEST_CHECK(len == sizeof(src) - 1))
+    {
+      TEST_MSG("Expected: %zu", sizeof(src) - 1);
+      TEST_MSG("Actual  : %zu", len);
+    }
+    if (!TEST_CHECK(strcmp(big, src) == 0))
+    {
+      TEST_MSG("Expected: %s", src);
+      TEST_MSG("Actual  : %s", big);
+    }
+  }
 }
